Bug #51105 ucs-test: Fix invalid Python raw strings
The backlash `\` is used to start an escape sequence:
<https://docs.python.org/3/reference/lexical_analysis.html#strings>
Using it before any other character leads to strange errors.

Convert all string to raw-strings (and fix regular expression)

And the real story is this:
    python3 -m flake8 tests/51_samba4/56evaluate_windows_gpo
> Process Process-1:
> Traceback (most recent call last):
...
> SyntaxError: (unicode error) 'unicodeescape' codec can't decode bytes in position 32-34: truncated \UXXXXXXXX escape
>
> During handling of the above exception, another exception occurred:
>
> Traceback (most recent call last):
...
> AttributeError: 'NoneType' object has no attribute 'rstrip'

This happens because since Python 3 unicode is the default, so
"\U........" is now a Python 3 string literal escape sequence, but
"\UCSTestKe" is not a valid unicode code point - only exactly 8 hex
chars are allowed and expected.
With Python 2 the "\U"-escape sequence only work in unicode literals,
but as strings are bytes there, the bug does not trigger with Python 2.