Bug #51105: Always use raw-strings for re.FUNC(r"...", ...)
The backlash `\` is used to start an escape sequence:
<https://docs.python.org/3/reference/lexical_analysis.html#strings>
Using it before any other character leads to strange errors.

The first argument to the 're' function is the pattern. As most often it
contains a backslash, using a r'...'-string is recommended. For
consistency always use raw-strings.

    git grep -n '\<re\.\w\+('[\'\"\]

FYI: Both re.sub('\n') and re.sub(r'\n') are equal: With the first the
re-compiler gets a string with the new-line already parsed by Python,
while with the second the backslash-sequence is only translated to
newline inside re.compile() itself. Quoting re:

> Most of the standard escapes supported by Python string literals are
> also accepted by the regular expression parser:
> \a \b \f \n \r \t \v \x \\