replication: Avoid expensive nested loop when loglevel below INFO
handler() goes into a nested for loop to figure out whether the
entry read from LDAP differs from the one it got from the cache.
The purpose of that seems to be to log a message (at "INFO" level)
when they differ.

That nested for loop can be every expensive. E.g. for very large
groups. (We ran into this with a group with roughly 100000 values
in the memberUid and uniqueMember attributes).

This patch skips the loop when the currently set loglevel is
below "INFO". It might make sense only go even further and to the
logs only and "DEBUG" level. (Or skip the checks if the number of
attribute values is beyond a certain size).